Fix compile error under msvc
git-svn-id: svn://svn.code.sf.net/p/qtvlm/code/branches/qtVlm_dev@1380 2387b160-4621-4e8a-be36-40ea10d72da2

+++ src/Dialogs/DialogRoutage.cpp
@@ -303,7 +303,7 @@ void DialogRoutage::done(int result)
         if(routage->getBoat()->getType()==BOAT_VLM)
         {
             time_t ddd=dd.toTime_t();
-            ddd=floor(ddd/routage->getBoat()->getVacLen())*routage->getBoat()->getVacLen();
+            ddd=floor((double)ddd/(double)routage->getBoat()->getVacLen())*routage->getBoat()->getVacLen();
             dd=dd.fromTime_t(ddd);
         }
         routage->setStartTime(dd);
